Moved goal screens to their own folder
