+++ code/TimeManagementApp/pages/goals/_tasks.js
@@ -22,8 +22,8 @@ import {
 
 import DateTimePicker from '@react-native-community/datetimepicker';
 import {NavigationContainer} from '@react-navigation/native';
-import { updateGoal, updateTask } from '../logic/goals.js';
-import { taskNotPresent, deleteTask } from '../logic/tasks.js';
+import { updateGoal, updateTask } from '../../logic/goals.js';
+import { taskNotPresent, deleteTask } from '../../logic/tasks.js';
 
 /**
   Returns the react components that are used to render the checklist
